LON IS NO MORE!!!
+ Added toml as the config load
+ Removed the Lua implementation
(I always hated .ini because it was ugly and too simple, because of that I started a Lua config manager, but that was consuming my soul. Things became too complex and I found out that Lua is not the besti option since it does not store table values ordered if they are indexed by strings, so I decided to find a good alternative to .ini, and after a quick search I found toml. No need to say that it does everything I wanted and much better than what I was writing)